Refactor to remove duplicated code.
Previously it first seeded the dataframe from the first in the list, then interated through the rest of the list.

All but the operator and the label was the same, therefore now account for when its the first iteration of the loop, and it switches the label to use on the dataframe.

For the statement building to switch between equals and not equals a new function created.